requirements.txt
torch
requests
watchdog
google-generativeai
python-dotenv
